refactor(platform): Actions minces → PlatformUserDirectoryService (Infrastructure) — conformité pureté des couches (Part of #6569, Closes #7004)
Les facades DB sont réservées à Interfaces/Infrastructure : la logique d'annuaire
(public.users, user_employee_links, companies, shared_tenants.employees) vit
désormais dans Infrastructure/Services/PlatformUserDirectoryService ; les Actions
Application délèguent (pattern ProvisionCompany → CompanyProvisioningService).
Comportement identique ; PlatformUsersApiTest inchangé.
